refactoring to support multidataset
