+++ examples/DynaCLR/embedding-web-visualization/interactive_visualizer.py
@@ -1,11 +1,12 @@
-"""Interactive visualization of phenotype data."""
-
 import logging
 from pathlib import Path
 
+import click
 from numpy.random import seed
 
-from viscy.representation.evaluation.visualization import EmbeddingVisualizationApp
+from viscy.representation.visualization.app import EmbeddingVisualizationApp
+from viscy.representation.visualization.settings import VizConfig
+from viscy.utils.cli_utils import yaml_to_model
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -13,33 +14,32 @@ logger = logging.getLogger(__name__)
 seed(42)
 
 
-def main():
+@click.command()
+@click.option(
+    "--config-filepath",
+    "-c",
+    required=True,
+    help="Path to YAML configuration file.",
+)
+def main(config_filepath):
     """Main function to run the visualization app."""
 
-    # Config for the visualization app
-    # TODO: Update the paths to the downloaded data. By default the data is downloaded to ~/data/dynaclr/demo
-    download_root = Path.home() / "data/dynaclr/demo"
-    output_path = Path.home() / "data/dynaclr/demo/embedding-web-visualization"
-    viz_config = {
-        "data_path": download_root / "registered_test.zarr",  # TODO add path to data
-        "tracks_path": download_root / "track_test.zarr",  # TODO add path to tracks
-        "features_path": download_root
-        / "precomputed_embeddings/infection_160patch_94ckpt_rev6_dynaclr.zarr",  # TODO add path to features
-        "channels_to_display": ["Phase3D", "RFP"],
-        "fov_tracks": {
-            "/A/3/9": list(range(50)),
-            "/B/4/9": list(range(50)),
-        },
-        "yx_patch_size": (160, 160),
-        "z_range": (24, 29),
-        "num_PC_components": 8,
-        "output_dir": output_path,
-    }
+    # Load and validate configuration from YAML file
+    viz_config = yaml_to_model(yaml_path=config_filepath, model=VizConfig)
+
+    output_dir = Path(__file__).parent / "output"
+    cache_path = (
+        Path("/home/eduardo.hirata/mydata/tmp/pcviewer") / "cache" / "image_cache.pkl"
+    )
 
     # Create and run the visualization app
     try:
-        # Create and run the visualization app
-        app = EmbeddingVisualizationApp(**viz_config)
+        app = EmbeddingVisualizationApp(
+            viz_config=viz_config,
+            cache_path=str(cache_path),
+            num_loading_workers=16,
+            output_dir=str(output_dir),
+        )
         app.preload_images()
         app.run(debug=True)
